LPS-99657 Move out of directory 'service'
