+++ modules/apps/organizations/organizations-service/src/main/java/com/liferay/organizations/internal/configuration/definition/OrganizationLinkConfigurationBeanDeclaration.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.organizations.service.internal.configuration.definition;
+package com.liferay.organizations.internal.configuration.definition;
 
 import com.liferay.organizations.configuration.OrganizationLinkConfiguration;
 import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
